mmc-test: add a simple r/w test of the sdmmc subsystem
The mmc_test.sh does the following steps:

Assumes an NFS filsystem with at least 2GB of free space.

- Creates 2GB test file of random using /dev/urandom. This test file is
  created on the NFS home directory, so this test will also test the
  ethernet data transfers.
- Writes the test file over NFS to the /dev/mmcblk0p4 SD partition.
- Compare the 2 files
- Cleanup by unmounting SD partition and deleting the test file.

This is an initial step towards just writing the SoCFPGA image to the
SD card and booting with that image. But before that can be done, we need
to define how that process will be done with the automated test system.

Signed-off-by: Dinh Nguyen <dinguyen@opensource.altera.com>
---
v2: address code review comments and add call to cleanup