spi-test-loopback: Update SPIDEV address.
Without the explicit bus number (removed in 3.16), spidev now
enumerates as 32766 instead of 0 (/dev/spidev32766.0). To
support both old and new kernels, this test was modified to
autodetect the bus number.

Signed-off-by: Thor Thayer <tthayer@opensource.altera.com>
---
v2: Update text to autodetect the bus number.

v3: Cleanup according to review comments. Addition of check
    for DevKit version. Addition of proper return code.

v4: Use return code of 0 for DevKits that don't support the
    loopback mode so LAVA is OK. Reorganize functions for a
    cleaner layout.
---
Tested on 3.10ltsi and 4.0 with and without shorting jumper
(no jumper means no loopback) on CycloneV. Also test Arria10
to ensure it exits cleanly.