Improvements to QSPI MTD tests
Only install kernel modules if
1) kernel module is not already installed,
2) kernel module is available

Since writing to wrong partition in Arria10 erases the U-Boot,
protect against this by checking which CPU is running using
DEV_PARTITION.

Add dump of dmesg for recording the mtd results.

Signed-off-by: Thor Thayer <thor.thayer@linux.intel.com>
---
v4  Replace all cases of - with _
    Use "$@" to grab all parameters in cleanup_modules()
v3  Change to a loop to install and then
    uninstall modules.
v2  Set QSPI partition to 1 for Stratix10.
    Wrap module path and arguments in quotes.
    Remove tail from dmesg
    Use modprobe -r to take care of dependencies.